fix print args in 64bit

+++ base/tests/ProcessInfo_test.cpp
@@ -15,9 +15,9 @@ std::string formatTime(std::chrono::system_clock::time_point tp) {
 	struct tm tm_time;
 	::localtime_r(&time, &tm_time);
 	char buf[32] = { 0 };
-	::snprintf(buf, sizeof(buf), "%4d%02d%02d %02d:%02d:%02d.%06lld",
+	::snprintf(buf, sizeof(buf), "%4d%02d%02d %02d:%02d:%02d.%06ld",
 		tm_time.tm_year + 1900, tm_time.tm_mon + 1, tm_time.tm_mday,
-		tm_time.tm_hour, tm_time.tm_min, tm_time.tm_sec, microseconds.count());
+		tm_time.tm_hour, tm_time.tm_min, tm_time.tm_sec, static_cast<int64_t>(microseconds.count()));
 	return buf;
 }
